+++ src/Zatomic.AI.Providers/HuggingFace/HuggingFaceChatRequest.cs
@@ -25,7 +25,7 @@ namespace Zatomic.AI.Providers.HuggingFace
 		public HuggingFaceChatBaseResponseFormat ResponseFormat { get; set; }
 
 		[JsonProperty("seed", NullValueHandling = NullValueHandling.Ignore)]
-		public int? Seed { get; set; }
+		public long? Seed { get; set; }
 
 		[JsonProperty("stop", NullValueHandling = NullValueHandling.Ignore)]
 		public List<string> Stop { get; set; }
